+++ src/ValueProvider/JsonFileValueProvider.mjs
@@ -2,7 +2,7 @@ import { FILE_SUFFIX } from "./FileValueProvider.mjs";
 import { JSON_SUFFIX } from "./JsonValueProvider.mjs";
 import { readFile } from "node:fs/promises";
 
-/** @typedef {import("../FluxConfigApi.mjs").FluxConfigApi} FluxConfigApi */
+/** @typedef {import("../FluxConfig.mjs").FluxConfig} FluxConfig */
 
 export const JSON_FILE_SUFFIX = `${JSON_SUFFIX}${FILE_SUFFIX}`;
 
@@ -23,15 +23,15 @@ export class JsonFileValueProvider {
 
     /**
      * @param {string} key
-     * @param {FluxConfigApi} flux_config_api
+     * @param {FluxConfig} flux_config
      * @returns {Promise<*>}
      */
-    async getConfig(key, flux_config_api) {
+    async getConfig(key, flux_config) {
         if (key.endsWith(FILE_SUFFIX) || key.endsWith(JSON_SUFFIX)) {
             return null;
         }
 
-        const value = await flux_config_api.getConfig(
+        const value = await flux_config.getConfig(
             `${key}${JSON_FILE_SUFFIX}`,
             null,
             false