+++ src/ValueProvider/JsonValueProvider.mjs
@@ -1,6 +1,6 @@
 import { FILE_SUFFIX } from "./FileValueProvider.mjs";
 
-/** @typedef {import("../FluxConfigApi.mjs").FluxConfigApi} FluxConfigApi */
+/** @typedef {import("../FluxConfig.mjs").FluxConfig} FluxConfig */
 
 export const JSON_SUFFIX = "-json";
 
@@ -21,15 +21,15 @@ export class JsonValueProvider {
 
     /**
      * @param {string} key
-     * @param {FluxConfigApi} flux_config_api
+     * @param {FluxConfig} flux_config
      * @returns {Promise<*>}
      */
-    async getConfig(key, flux_config_api) {
+    async getConfig(key, flux_config) {
         if (key.endsWith(FILE_SUFFIX) || key.endsWith(JSON_SUFFIX)) {
             return null;
         }
 
-        const value = await flux_config_api.getConfig(
+        const value = await flux_config.getConfig(
             `${key}${JSON_SUFFIX}`,
             null,
             false